+++ src/extension.ts
@@ -1,7 +1,6 @@
 import * as vscode from "vscode";
 
 import { sortImports } from "./utils/sortImports";
-import { sortImportsPreservingComments } from "./utils/sortImportsPreservingComments";
 
 export function activate(context: vscode.ExtensionContext) {
   vscode.commands.registerCommand("customImportSort.sortImports", () => {
@@ -33,39 +32,6 @@ export function activate(context: vscode.ExtensionContext) {
       }
     }
   });
-
-  vscode.commands.registerCommand(
-    "customImportSort.sortImportsPreservingComments",
-    () => {
-      try {
-        const editor = vscode.window.activeTextEditor;
-        if (editor) {
-          const document = editor.document;
-          const fileName = document.fileName;
-          const extension = fileName.substring(fileName.lastIndexOf(".") + 1);
-          if (/^[jt]sx?$/.test(extension)) {
-            const text = document.getText();
-            const sortedText = sortImportsPreservingComments(text);
-            if (text !== sortedText) {
-              editor.edit((editBuilder) => {
-                editBuilder.replace(
-                  new vscode.Range(
-                    new vscode.Position(0, 0),
-                    new vscode.Position(document.lineCount, 0)
-                  ),
-                  sortedText
-                );
-              });
-            }
-          }
-        }
-      } catch (error) {
-        if (error instanceof Error) {
-          vscode.window.showErrorMessage(error.message);
-        }
-      }
-    }
-  );
 
   // run customImportSort.sortImports on save if enabled
   context.subscriptions.push(